Add files via upload
In Java, the LinkedList class represents a dynamic, ordered set of objects. It inherits the List interface and part of the implementation of the AbstractSequentialList class.
You can add elements to a LinkedList using the add() method. The add() method adds elements to the end of the list.
You can sort a LinkedList in Java using the sort() method. The sort() method sorts the elements in the list in their natural ascending order.